Renamed admin directory to admin_d directory
This was done to avoid the name clash between the admin module (auto-discovered
and improted by django.contrib.admin) and the admin directory with all the
code for the admin module.
