+++ djstripe/admin_d/admin_inline.py
